requirements.txt
aiogram==3.0.0b7
SQLAlchemy==1.4.52
asyncpg==0.27.0
alembic==1.11.1
python-dotenv==1.0.0
